Moving tree branches now updates cached full_names correctly.
Because we cache a bit of information on each AbstractTopic node (the
full_name attribute), this must be updated when a subtree is moved.
Since subclasses of AbstractTopic may have similar requirements, there's
also a facility to pass in other callables which are called for each
child node to do similar updates prior to saving the node.